fix: seperate into files to fix the css import error
