--- src/components/EmailTemplate.module.css
@@ -1,11 +1,0 @@
-.defaultBodyStyle {
-  display: flex;
-  flex-direction: column;
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  height: 100vh;
-  align-items: center;
-  justify-content: center;
-  overflow: auto;
-}
