+++ src/components/EmailTemplateClient.tsx
@@ -3,8 +3,7 @@
 // organize-imports-ignore
 import React from 'react'
 import type { RenderEmailTemplateProps } from '../utils/renderEmailTempalte.js'
-import styles from './EmailTemplate.module.css'
-import { BlockRendererClient } from './BlockRendererFactory.js'
+import { BlockRendererClient } from './BlockRenderer/BlockRendererClient.js'
 import { Block } from '../types.js'
 import { EmailTemplatePlaceholder } from './EmailTemplatePlaceholder.js'
 import type { UploadCollectionSlug } from 'payload'
@@ -18,8 +17,26 @@ export const EmailTemplateClient = (
 
   const body = data.body
 
+  const defaultStyle = {
+    display: 'flex',
+    flexDirection: 'column',
+    margin: 0,
+    padding: 0,
+    boxSizing: 'border-box',
+    height: '100vh',
+    alignItems: 'center',
+    justifyContent: 'center',
+    overflow: 'auto',
+  }
+
   return (
-    <div id="email-template-client" className={styles.defaultBodyStyle} style={data.style}>
+    <div
+      id="email-template-client"
+      style={{
+        ...defaultStyle,
+        ...data.style,
+      }}
+    >
       {body && Array.isArray(body) && body.length > 0 ? (
         body.map((block: Block) =>
           BlockRendererClient({ block, previewMode: 'preview', imageCollectionSlug }),
